refactring connecting code screen
