+++ lib/common/constants/image_assets.dart
@@ -1,0 +1,4 @@
+abstract class ImageAssets {
+  static const splash = 'assets/img/bg_splash.png';
+  static const document = 'assets/icons/what_get_icon.svg';
+}
