+++ lib/feature/presentation/navigation_route_names.dart
@@ -1,13 +1,14 @@
 import 'package:cportal_flutter/feature/domain/entities/onboarding_entity.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/news_bloc/news_code_enum.dart';
-import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connection_code_web/connecting_code_web.dart';
-import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connection_code_web/qr_scanner_web.dart';
-import 'package:cportal_flutter/feature/presentation/ui/connecting_code/qr_scanner.dart';
+import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connecting_code_mobile/connecting_code_info_popup.dart';
+import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connecting_code_mobile/qr_scanner.dart';
+import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connecting_code_screen.dart';
+import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connection_code_web/connecting_code_info_screen.dart';
+import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connection_code_web/connecting_qr_screen.dart';
 import 'package:cportal_flutter/feature/presentation/ui/contacts_page/contact_profile_page.dart';
 import 'package:cportal_flutter/feature/presentation/ui/contacts_page/contacts_page.dart';
 import 'package:cportal_flutter/feature/presentation/ui/declarations_page/declarations_page.dart';
 import 'package:cportal_flutter/feature/presentation/ui/home/home_page.dart';
-import 'package:cportal_flutter/feature/presentation/ui/connecting_code/connecting_code_page.dart';
 import 'package:cportal_flutter/feature/presentation/ui/finger_print/finger_print_or_faceid_page.dart';
 import 'package:cportal_flutter/feature/presentation/ui/main_page/main_page.dart';
 import 'package:cportal_flutter/feature/presentation/ui/news_page/articles/news_article_page.dart';
@@ -32,7 +33,10 @@ abstract class NavigationRouteNames {
   static const splashScreen = 'splash_screen';
   static const mainPage = 'main_page';
   static const connectingCode = 'connecting_code';
-  static const connectingCodeWeb = 'connecting_code_web';
+  static const connectingCodeInfoPopup = 'connecting_code_info_popup';
+  static const connectingCodeInfo = 'connecting_code_info';
+  static const connectingQr = 'connecting_qr';
+  static const qrScanner = 'qr_scanner';
   static const createPin = 'create_pin';
   static const createPinWeb = 'create_pin_web';
   static const inputPin = 'input_pin';
@@ -47,8 +51,6 @@ abstract class NavigationRouteNames {
   static const questionArticlePage = 'question_article_page';
   static const profile = 'profile';
   static const userData = 'user_data';
-  static const qrScanner = 'qr_scanner';
-  static const qrScannerWeb = 'qr_scanner_web';
   static const onBoardingStart = 'onboarding_start';
   static const onboarding = 'onboarding';
   static const onboardingEnd = 'onboarding_end';
@@ -59,9 +61,7 @@ abstract class NavigationRouteNames {
 
 final GoRouter router = GoRouter(
   urlPathStrategy: UrlPathStrategy.path,
-  initialLocation: '/',
-  // ignore: format-comment
-  // debugLogDiagnostics: true,
+  initialLocation: '/splash_screen',
   routes: <GoRoute>[
     GoRoute(
       name: NavigationRouteNames.splashScreen,
@@ -69,6 +69,59 @@ final GoRouter router = GoRouter(
       pageBuilder: (context, state) => MaterialPage(
         key: state.pageKey,
         child: const Scaffold(body: SplashScreen()),
+      ),
+    ),
+    GoRoute(
+      name: NavigationRouteNames.connectingCode,
+      path: '/connecting_code',
+      pageBuilder: (context, state) => MaterialPage(
+        key: state.pageKey,
+        child: const ConnectingCodeScreen(),
+      ),
+      routes: [
+        GoRoute(
+          name: NavigationRouteNames.connectingCodeInfoPopup,
+          path: 'info_popup',
+          pageBuilder: (context, state) => CustomTransitionPage(
+            key: state.pageKey,
+            barrierColor: Colors.black54,
+            barrierDismissible: true,
+            fullscreenDialog: true,
+            opaque: false,
+            child: const ConnectingCodeInfoPopup(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) => FadeTransition(
+              opacity: CurvedAnimation(
+                parent: animation,
+                curve: Curves.easeOut,
+              ),
+              child: child,
+            ),
+          ),
+        ),
+        GoRoute(
+          name: NavigationRouteNames.connectingCodeInfo,
+          path: 'info',
+          pageBuilder: (context, state) => MaterialPage(
+            key: state.pageKey,
+            child: const ConnectingCodeInfoScreen(),
+          ),
+        ),
+      ],
+    ),
+    GoRoute(
+      name: NavigationRouteNames.qrScanner,
+      path: '/qr_scanner',
+      pageBuilder: (context, state) => MaterialPage(
+        key: state.pageKey,
+        child: const QrScanner(),
+      ),
+    ),
+    GoRoute(
+      name: NavigationRouteNames.connectingQr,
+      path: '/qr_scanner_web',
+      pageBuilder: (context, state) => MaterialPage(
+        key: state.pageKey,
+        child: const ConnectingQrScreen(),
       ),
     ),
     GoRoute(
@@ -80,22 +133,6 @@ final GoRouter router = GoRouter(
           child: MainPage(),
           desktopMenuIndex: 0,
         ),
-      ),
-    ),
-    GoRoute(
-      name: NavigationRouteNames.connectingCode,
-      path: '/connecting_code',
-      pageBuilder: (context, state) => MaterialPage(
-        key: state.pageKey,
-        child: const ConnectingCodePage(),
-      ),
-    ),
-    GoRoute(
-      name: NavigationRouteNames.connectingCodeWeb,
-      path: '/connecting_code_web',
-      pageBuilder: (context, state) => MaterialPage(
-        key: state.pageKey,
-        child: const Scaffold(body: ConnectingCodeWeb()),
       ),
     ),
     GoRoute(
@@ -241,22 +278,6 @@ final GoRouter router = GoRouter(
       pageBuilder: (context, state) => MaterialPage(
         key: state.pageKey,
         child: const OnBoardingWelcome(),
-      ),
-    ),
-    GoRoute(
-      name: NavigationRouteNames.qrScanner,
-      path: '/qr_scanner',
-      pageBuilder: (context, state) => MaterialPage(
-        key: state.pageKey,
-        child: const QrScanner(),
-      ),
-    ),
-    GoRoute(
-      name: NavigationRouteNames.qrScannerWeb,
-      path: '/qr_scanner_web',
-      pageBuilder: (context, state) => MaterialPage(
-        key: state.pageKey,
-        child: const QrScannerWeb(),
       ),
     ),
     GoRoute(
