+++ lib/feature/presentation/ui/connecting_code/widgets/what_get_with_you.dart
@@ -15,42 +15,7 @@ class WhatGetWithYou extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-                                     final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
-
-
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        SvgPicture.asset(
-          iconPath,
-          width: 24,
-          color: color,
-        ),
-        const SizedBox(height: 14.67),
-        Text(
-          text,
-          style: theme.textTheme.px16,
-        ),
-      ],
-    );
-  }
-}
-
-class WhatGetWithYouWeb extends StatelessWidget {
-  final String iconPath;
-  final Color? color;
-  final String text;
-  const WhatGetWithYouWeb({
-    Key? key,
-    required this.iconPath,
-    this.color,
-    required this.text,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-                                  final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
-
+    final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
 
     return Column(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
