+++ lib/feature/presentation/ui/connecting_code/widgets/work_mode_table.dart
@@ -7,8 +7,7 @@ class WorkModeTable extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-                                     final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
-
+    final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
     final locale = AppLocalizations.of(context)!;
 
     final workModeMap = [
