+++ lib/feature/presentation/ui/splash_screen/splash_screen.dart
@@ -2,95 +2,28 @@ import 'package:cportal_flutter/feature/presentation/bloc/auth_bloc/auth_bloc.da
 import 'package:cportal_flutter/feature/presentation/bloc/auth_bloc/auth_event.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/auth_bloc/auth_state.dart';
 import 'package:cportal_flutter/feature/presentation/navigation_route_names.dart';
-import 'package:cportal_flutter/feature/presentation/ui/main_page/widgets/svg_icon.dart';
+import 'package:cportal_flutter/feature/presentation/ui/widgets/splash_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 
 class SplashScreen extends StatelessWidget {
-  static const String routeName = 'splashScreen';
-
   const SplashScreen({Key? key}) : super(key: key);
-  static Route<SplashScreen> route() {
-    return MaterialPageRoute<SplashScreen>(
-      settings: const RouteSettings(name: routeName),
-      builder: (context) => const SplashScreen(),
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
-    BlocProvider.of<AuthBloc>(context, listen: false).add(const CheckAuth());
+    context.read<AuthBloc>().add(const CheckAuth());
 
-    return BlocListener<AuthBloc, AuthState>(
-      listener: (context, state) {
-        if (state is Authenticated) {
-          final nextScreen = state.isAuth
-              ? NavigationRouteNames.mainPage
-              : NavigationRouteNames.connectingCode;
+    return Scaffold(
+      body: BlocListener<AuthBloc, AuthState>(
+        listener: (context, state) {
+          if (state is Authenticated) {
+            final nextScreen = state.isAuth ? NavigationRouteNames.mainPage : NavigationRouteNames.connectingCode;
 
-          context.goNamed(nextScreen);
-        }
-      },
-      child: const LoaderWidget(),
-    );
-  }
-}
-
-class LoaderWidget extends StatelessWidget {
-  const LoaderWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: const BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage('assets/img/bg_splash.png'),
-          fit: BoxFit.cover,
-          colorFilter: ColorFilter.mode(
-            Color.fromRGBO(37, 39, 40, 0.7),
-            BlendMode.darken,
-          ),
-        ),
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: const[
-          Center(
-            child: SvgIcon(null, path: 'logo.svg', width: 120),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class LoaderWebWidget extends StatelessWidget {
-  const LoaderWebWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final double width = MediaQuery.of(context).size.width;
-
-    return Container(
-      width: width * 0.42,
-      decoration: const BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage('assets/img/bg_splash.png'),
-          fit: BoxFit.cover,
-          colorFilter: ColorFilter.mode(
-            Color.fromRGBO(37, 39, 40, 0.7),
-            BlendMode.darken,
-          ),
-        ),
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: const [
-          Center(
-            child: SvgIcon(null, path: 'logo.svg', width: 120),
-          ),
-        ],
+            context.goNamed(nextScreen);
+          }
+        },
+        child: const SplashWidget.mobile(),
       ),
     );
   }
